requirements.txt
# ===== ETAPA 1: Knowledge Base =====
pandas==2.2.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# ===== ETAPA 2: RAG Pipeline =====
langchain==0.1.20
langchain-community==0.0.38
langchain-text-splitters==0.0.1
chromadb==0.4.24
sentence-transformers==2.5.1

# ===== ETAPA 2: LLM (Gemini) =====
google-generativeai==0.4.1